Add types, refactor code
